Update functions.php
Hi, I'm using your wp plugin and I've seen a mistake on this line. When is a weekend or holiday '$is_holidy' should be true, but your comparison return false. I think that is because the type of '$work_days[$key]' is not an integer like '0'. I changed it with a simple equal comparison and it worked ..

I hope that this help you.

Best regards!